+++ toolkit/components/downloads/DownloadCore.sys.mjs
@@ -45,6 +45,13 @@ const BackgroundFileSaverStreamListener = Components.Constructor(
   "nsIBackgroundFileSaver"
 );
 
+XPCOMUtils.defineLazyPreferenceGetter(
+  lazy,
+  "enableDeletePrivateFeature",
+  "browser.download.enableDeletePrivate",
+  false
+);
+
 /**
  * Returns true if the given value is a primitive string or a String object.
  */
@@ -87,9 +94,7 @@ function deserializeUnknownProperties(aObject, aSerializable, aFilterFn) {
 /**
  * Check if the file is a placeholder.
  *
- * @return {Promise}
- * @resolves {boolean}
- * @rejects Never.
+ * @returns {Promise<boolean>}
  */
 async function isPlaceholder(path) {
   try {
@@ -203,7 +208,7 @@ Download.prototype = {
    * Progress percent, from 0 to 100.  Intermediate values are reported only if
    * hasProgress is true.
    *
-   * @note You shouldn't rely on this property being equal to 100 to determine
+   * Note: You shouldn't rely on this property being equal to 100 to determine
    *       whether the download is completed.  You should use the individual
    *       state properties instead.
    */
@@ -215,7 +220,7 @@ Download.prototype = {
    *
    * When hasProgress is false, this property is always zero.
    *
-   * @note This property may be different than the final file size on disk for
+   * Note: This property may be different than the final file size on disk for
    *       downloads that are encoded during the network transfer.  You can use
    *       the "size" property of the DownloadTarget object to get the actual
    *       size on disk once the download succeeds.
@@ -226,7 +231,7 @@ Download.prototype = {
    * Number of bytes currently transferred.  This value starts at zero, and may
    * be updated regardless of the value of hasProgress.
    *
-   * @note You shouldn't rely on this property being equal to totalBytes to
+   * Note: You shouldn't rely on this property being equal to totalBytes to
    *       determine whether the download is completed.  You should use the
    *       individual state properties instead.  This property may not be
    *       updated during the last part of the download.
@@ -362,12 +367,12 @@ Download.prototype = {
    * finished yet, this method waits for the cancellation to finish, then
    * restarts the download immediately.
    *
-   * @note If you need to start a new download from the same source, rather than
+   * Note: If you need to start a new download from the same source, rather than
    *       restarting a failed or canceled one, you should create a separate
    *       Download object with the same source as the current one.
    *
-   * @return {Promise}
-   * @resolves When the download has finished successfully.
+   * @returns {Promise<void>}
+   *   Resolves when the download has finished successfully.
    * @rejects JavaScript exception if the download failed.
    */
   start: function D_start() {
@@ -631,8 +636,8 @@ Download.prototype = {
   /**
    * Perform the actions necessary when a Download succeeds.
    *
-   * @return {Promise}
-   * @resolves When the steps to take after success have completed.
+   * @returns {Promise<void>}
+   *   Resolves when the steps to take after success have completed.
    * @rejects  JavaScript exception if any of the operations failed.
    */
   async _succeed() {
@@ -643,11 +648,20 @@ Download.prototype = {
     if (this.launchWhenSucceeded) {
       this.launch().catch(console.error);
 
-      if (this.source.isPrivate) {
+      // If the "have users choose whether to delete private downloads" feature
+      // is not enabled, we should delete the temporary file once the private
+      // browsing session ends. Otherwise, we'll decide whether to delete the
+      // file further down, for all downloads (not just the ones that have
+      // `launchWhenSucceeded` set to true).
+      // Once we complete rollout of `enableDeletePrivateFeature`, we can
+      // remove this block.
+      if (!lazy.enableDeletePrivateFeature && this.source.isPrivate) {
         lazy.gExternalAppLauncher.deleteTemporaryPrivateFileWhenPossible(
           new lazy.FileUtils.File(this.target.path)
         );
-      } else if (
+      }
+      if (
+        !this.source.isPrivate &&
         Services.prefs.getBoolPref("browser.helperApps.deleteTempFileOnExit") &&
         Services.prefs.getBoolPref(
           "browser.download.start_downloads_in_tmp_dir",
@@ -661,10 +675,7 @@ Download.prototype = {
     }
 
     if (
-      Services.prefs.getBoolPref(
-        "browser.download.enableDeletePrivate",
-        false
-      ) &&
+      lazy.enableDeletePrivateFeature &&
       Services.prefs.getBoolPref("browser.download.deletePrivate", false) &&
       this.source.isPrivate
     ) {
@@ -695,8 +706,8 @@ Download.prototype = {
    * The file will be moved out of quarantine and the download will be
    * marked as succeeded.
    *
-   * @return {Promise}
-   * @resolves When the Download has been unblocked and succeeded.
+   * @returns {Promise<void>}
+   *   Resolves when the Download has been unblocked and succeeded.
    * @rejects  JavaScript exception if any of the operations failed.
    */
   unblock() {
@@ -826,8 +837,8 @@ Download.prototype = {
    * If a download was blocked but retained on disk this method can be used
    * to remove the file.
    *
-   * @return {Promise}
-   * @resolves When the Download's data has been removed.
+   * @returns {Promise<void>}
+   *   Resolves when the Download's data has been removed.
    * @rejects  JavaScript exception if any of the operations failed.
    */
   confirmBlock() {
@@ -874,7 +885,7 @@ Download.prototype = {
     return this._promiseConfirmBlock;
   },
 
-  /*
+  /**
    * Launches the file after download has completed. This can open
    * the file with the default application for the target MIME type
    * or file extension, or with a custom application if launcherPath
@@ -886,8 +897,8 @@ Download.prototype = {
    * @param options.useSystemDefault
    *                           Optional value indicating how to handle launching this download,
    *                           this time only. Will override the associated mimeInfo.preferredAction
-   * @return {Promise}
-   * @resolves When the instruction to launch the file has been
+   * @returns {Promise<void>}
+   *   Resolves when the instruction to launch the file has been
    *           successfully given to the operating system. Note that
    *           the OS might still take a while until the file is actually
    *           launched.
@@ -908,13 +919,13 @@ Download.prototype = {
     return lazy.DownloadIntegration.launchDownload(this, options);
   },
 
-  /*
+  /**
    * Shows the folder containing the target file, or where the target file
    * will be saved. This may be called at any time, even if the download
    * failed or is currently in progress.
    *
-   * @return {Promise}
-   * @resolves When the instruction to open the containing folder has been
+   * @returns {Promise<void>}
+   *   Resolves when the instruction to open the containing folder has been
    *           successfully given to the operating system. Note that
    *           the OS might still take a while until the folder is actually
    *           opened.
@@ -956,8 +967,8 @@ Download.prototype = {
    * to "cancel" happens asynchronously, and is consistent with the case where
    * the cancellation request could not be processed in time.
    *
-   * @return {Promise}
-   * @resolves When the cancellation process has finished.
+   * @returns {Promise<void>}
+   *   Resolves when the cancellation process has finished.
    * @rejects Never.
    */
   cancel: function D_cancel() {
@@ -1030,8 +1041,8 @@ Download.prototype = {
    * still true when the download is restarted, partial data will be retained
    * during the new download attempt.
    *
-   * @return {Promise}
-   * @resolves When the partial data has been successfully removed.
+   * @returns {Promise<void>}
+   *   Resolves when the partial data has been successfully removed.
    * @rejects JavaScript exception if the operation could not be completed.
    */
   removePartialData() {
@@ -1099,8 +1110,8 @@ Download.prototype = {
    * download is not controlled interactively, you should use the promise
    * returned by the "start" method instead, to check for success or failure.
    *
-   * @return {Promise}
-   * @resolves When the download has finished successfully.
+   * @returns {Promise<void>}
+   *   Resolves when the download has finished successfully.
    * @rejects Never.
    */
   whenSucceeded: function D_whenSucceeded() {
@@ -1116,8 +1127,8 @@ Download.prototype = {
    * This allows the properties of the download to be updated in case the user
    * moved or deleted the target file or its associated ".part" file.
    *
-   * @return {Promise}
-   * @resolves When the operation has completed.
+   * @returns {Promise<void>}
+   *   Resolves when the operation has completed.
    * @rejects Never.
    */
   refresh() {
@@ -1209,8 +1220,8 @@ Download.prototype = {
    *        Whether any partially downloaded data should be removed after the
    *        download has been stopped.
    *
-   * @return {Promise}
-   * @resolves When the operation has finished successfully.
+   * @returns {Promise<void>}
+   *   Resolves when the operation has finished successfully.
    * @rejects JavaScript exception if an error occurred while removing the
    *          partially downloaded data.
    */
@@ -1580,14 +1591,14 @@ DownloadSource.prototype = {
    * can adjust the network channel before it is opened, for example to change
    * the HTTP headers or to upload a stream as POST data.
    *
-   * @note If this is defined this object will not be serializable, thus the
+   * Note: If this is defined this object will not be serializable, thus the
    *       Download object will not be persisted across sessions.
    *
    * @param aChannel
    *        The nsIChannel to be adjusted.
    *
-   * @return {Promise}
-   * @resolves When the channel has been adjusted and can be opened.
+   * @returns {Promise<void>}
+   *   Resolves when the channel has been adjusted and can be opened.
    * @rejects JavaScript exception that will cause the download to fail.
    */
   adjustChannel: null,
@@ -1597,7 +1608,7 @@ DownloadSource.prototype = {
    * will determine, if provided, if a download can progress or has to be
    * cancelled based on the HTTP status code of the network channel.
    *
-   * @note If this is defined this object will not be serializable, thus the
+   * Note: If this is defined this object will not be serializable, thus the
    *       Download object will not be persisted across sessions.
    *
    * @param aDownload
@@ -1605,7 +1616,7 @@ DownloadSource.prototype = {
    * @param aStatus
    *        The HTTP status in question
    *
-   * @return {Boolean} Download can progress
+   * @return {boolean} Download can progress
    */
   allowHttpStatus: null,
 
@@ -1845,8 +1856,8 @@ DownloadTarget.prototype = {
   /**
    * Sets the "exists" and "size" properties based on the actual file on disk.
    *
-   * @return {Promise}
-   * @resolves When the operation has finished successfully.
+   * @returns {Promise<void>}
+   *   Resolves when the operation has finished successfully.
    * @rejects JavaScript exception.
    */
   async refresh() {
@@ -2020,7 +2031,7 @@ export var DownloadError = function (aProperties) {
  * These constants are used by the reputationCheckVerdict property and indicate
  * the detailed reason why a download is blocked.
  *
- * @note These values should not be changed because they can be serialized.
+ * Note: These values should not be changed because they can be serialized.
  */
 DownloadError.BLOCK_VERDICT_MALWARE = "Malware";
 DownloadError.BLOCK_VERDICT_POTENTIALLY_UNWANTED = "PotentiallyUnwanted";
@@ -2174,8 +2185,8 @@ DownloadSaver.prototype = {
    *        the names of the properties to set, and the value represents the
    *        value to set.
    *
-   * @return {Promise}
-   * @resolves When the download has finished successfully.
+   * @returns {Promise<void>}
+   *   Resolves when the download has finished successfully.
    * @rejects JavaScript exception if the download failed.
    */
   async execute() {
@@ -2199,8 +2210,8 @@ DownloadSaver.prototype = {
    *
    * @param canRemoveFinalTarget
    *        True if can remove target file regardless of it being a placeholder.
-   * @return {Promise}
-   * @resolves When the operation has finished successfully.
+   * @returns {Promise<void>}
+   *   Resolves when the operation has finished successfully.
    * @rejects Never.
    */
   async removeData() {},
@@ -2704,8 +2715,8 @@ DownloadCopySaver.prototype = {
    * @param aSetPropertiesFn
    *        Function provided to the "execute" method.
    *
-   * @return {Promise}
-   * @resolves When the reputation check and cleanup is complete.
+   * @returns {Promise<void>}
+   *   Resolves when the reputation check and cleanup is complete.
    * @rejects DownloadError if the download should be blocked.
    */
   async _checkReputationAndMove(aSetPropertiesFn) {
